jarvis_assistant/requirements.txt
SpeechRecognition
wyoming
wyoming
google-generativeai
openai
python-dotenv
pvporcupine
piper-tts
